Allow for configurable attribute transformer for Nazrin searchable struct
